requirements.txt
# python 3.8.5
pandas==1.5.2
numpy==1.24.1
scikit-learn==1.2.0
torch==1.13.0
transformers==4.25.1
wandb==0.13.7
black==22.10.0
datasets==2.8.0
fastapi==0.89.1
ffmpeg==1.4
flake8==5.0.4
git-lfs==1.6
isort==5.10.1
konlpy==0.6.0
mecab-python==0.996-ko-0.9.2
omegaconf==2.3.0
openai-whisper==20230124
pip-chill==1.0.1
pre-commit==2.17.0
pytest==7.0.1
python-mecab-ko==1.3.3
python-multipart==0.0.5
rank-bm25==0.2.2
rouge==1.0.1
rouge-score==0.1.2
sentence-transformers==2.2.2
streamlit-nested-layout==0.1.1
streamlit_tags==1.2.8
utils==1.0.1
uvicorn==0.20.0
